Add argument parsing.

+++ include/genome_index.hpp
@@ -83,7 +83,7 @@ namespace genomics {
 
         if (k >= mismatches) return;
 
-        for (size_t i = 0; i < search_alphabet_size; i++) {
+        for (int i = 0; i < search_alphabet_size; i++) {
             if (search_alphabet[i] == c) continue;
 
             char a = search_alphabet[i];
